backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# MongoDB
pymongo

# K2 Think V2
httpx==0.28.1

# GCP Vertex AI (Imagen)
google-auth==2.38.0

# Vector DB
chromadb==0.5.5
